# swagger.yaml
basePath: /
consumes:
    - application/json
...
info:
    description: '# Documentation for Product API'
    title: of Product API
    version: 1.0.0
paths:
...
    /products/{id}:
        delete:
            description: Returns a list of pr
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
produces:
    - application/json
responses:
    noContent:
        description: ""
...
schemes:
    - http
swagger: "2.0"